requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.3
Flask-Cors==3.0.
Werkzeug>2.0.0

